+++ webex_assistant_skills_sdk/shared/services/settings.py
@@ -7,6 +7,7 @@ from pydantic import BaseSettings
 class Settings(BaseSettings):
     skill_name: Optional[str]
     private_key_path: Path = 'id_rsa.pem'
+    public_key_path: Path = 'id_rsa.pub'
     secret: Optional[str]
     use_encryption: bool = True
     log_level: str = 'INFO'
